+++ .travis.yml
@@ -18,6 +18,7 @@ matrix:
         - php: 7.1
         - php: 7.2
         - php: 7.3
+          env: SYMFONY_PHPUNIT_VERSION=7.2
         - php: nightly
     allow_failures:
         - php: nightly
